Implement gpt-oss 20b and tinychat interface with nearly identical results to the reference implementation, though that comes at a cost because it is using float32 in some places because that is how some of pytorch works.  The one remaining thing I couldn't quite make 100% identical was einsum still has slight differences even with float32, but like 1e-7 to 1e-5, though that accumulates several times per layer and over each layer, leading to very slightly different generations sometimes.
Around 13 tok/s, using ~45gb of vram, including a kv cache.

For an idea of how accurate it is, we get the similar output for first 30 tokens
in response to 'why?'. The first 16 are identical and then error must have
accumulated enough to be slightly different (though still good) with
temperature 0

Still need to download weights and place it somewhere then use model
param.

Started to look at what it will take to use fp4, that is probably one of
the big keys to speedup, but it is very incomplete.

+++ test/test_linearizer.py
@@ -309,7 +309,7 @@ class TestLinearizer(unittest.TestCase):
         a, b = Tensor.rand(8, 8, dtype=tensor_dtype), Tensor.rand(8, 8, dtype=tensor_dtype)
         helper_arg_acc_dtype(a.sum(dtype=acc_dtype), expected_dtype)
         helper_arg_acc_dtype(a.matmul(b, dtype=acc_dtype), expected_dtype)
-        helper_arg_acc_dtype(Tensor.einsum("ki,ij->kj", a, b, dtype=acc_dtype), expected_dtype)
+        helper_arg_acc_dtype(Tensor.einsum("ki,ij->kj", a, b, math_dtype=acc_dtype), expected_dtype)
         d, w = Tensor.rand(4, 8, 8, 8, dtype=tensor_dtype), Tensor.rand(8, 8, 2, 2, dtype=tensor_dtype)
         helper_arg_acc_dtype(d.conv2d(w, dtype=acc_dtype), expected_dtype)
